ci: use ruff
This commit begins the first in what will hopefully be a short set of
continuous integration commits, adding ruff checking to the repo. On
pushes and pull requests, the ruff action will run. This checks the code
against the rules specified in ruff.toml and fixes all minor issues. It
also reports any errors and other issues for manual fixing.

This will ensure that all code pushed to this repository remains on the
same page, as it will all be reformatted according to the listed rules.
This should cut down on the number of code formatting issues and commits
necessary to fix those.