fix: handle NotImplemented error in fft angle computation
Complex angle calculations in torch tensors are not yet implemented for
metal processing shaders (MPS; Apple's GPU language). However, we can
still take advantage of gpu acceleration for all other calculations. To
do this, we initially try to do the calculation on the specified device.
If it is not implemented, then the code will fall back to running the
calculation on the cpu.

I chose to use a try/except block with NotImplemented to anticipate
other potential gpu languages that might not support the angle
calculations. This will hopefully also allow users to immediately take
advantage of the speed boost if torch supports mps angle calculations in
the future.

An alternate way to do this would be to specifically move these
calculations to the cpu only if the specified device is MPS. I'm not
sure that this way would be better, so I've set it up like so for now.

Happy coding!