feat!: BREAKING CHANGE partially support mps cwt acceleration
Although ssqueezepy does not support gpu acceleration on non-cuda
devices, some of the remaining calculations can still be sped up by
running them on the gpu.

This commit modifies the `query_cwt_block()` to allow non-ssqueezepy
calculations to run on the specified device by replacing the `use_gpu`
argument with a `device` argument that takes a torch device. I feel like
this makes sense as it brings the parameters for these functions in line
with the cellstream fft module, where all the functions take a device
argument.

BREAKING CHANGE! This commit replaces an argument for the following
functions:
* query_cwt_block(): use_gpu (bool) -> device (torch.device)
* generate_cwt_image_cellstreams(): use_gpu (bool) -> device
  (torch.device)